Added htmlspecialchars to views and fixed Static referencing
While some templating is preferred the current views generate the html
markup for printing the list of issues to the screen.  Since Summary and
Description are the two places where HTML lives I have sanitized that
for now as a cheap fix for HTML in those two fields breaking things.

fixes: #5